correct typo error

+++ scripts/destroy.sh
@@ -73,7 +73,7 @@ if [ ! -f "../backend/lambda-deployment.zip" ]; then
 fi
 
 # Run terraform destroy with auto-approve
-iif [ "$ENVIRONMENT" = "prod" ] && [ -f "prod.tfvars" ]; then
+if [ "$ENVIRONMENT" = "prod" ] && [ -f "prod.tfvars" ]; then
     TF_VAR_openai_api_key="${TF_VAR_openai_api_key}" terraform destroy -var-file=prod.tfvars -var="project_name=$PROJECT_NAME" -var="environment=$ENVIRONMENT" -auto-approve
 else
     TF_VAR_openai_api_key="${TF_VAR_openai_api_key}" terraform destroy -var="project_name=$PROJECT_NAME" -var="environment=$ENVIRONMENT" -auto-approve
